feat(module): Support for Native Element in both Target and Container (#56)
Native Elements are now suported as selectors for both `target` and `container` properties. This allows for a bit more finegrained control as to what is send into the `scrollTo` method. Users can now get elements using the ubiquitous `document.getElementById('#id')` and pass a reference to that value into the `scrollTo` method.

This closes #22 
This closes #55